refactor /transcode route
